Compute the Next N Samples
This should work the same as the last one.
I'm not actually using it for anything yet...

I now have the ability to compute the next N values, and the default is just to
grab the first.

If I stopped here, then this is just a more complicated and expensive way to do
the same thing I was doing before.

I hope to use this shortly, though.